requirements.txt
pillow
numpy
matplotlib
tensorflow
scikit-image
scipy